[HM-16]: Refractory component code
